Correct (int) cast(getUserByID and etc.), wrong pathes by add, reg and delete form

+++ src/com/bsu/by/komodey/controllers/InfoServlet.java
@@ -18,7 +18,7 @@ public class InfoServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        long id = Long.parseLong(req.getParameter("id"));
+        int id = Integer.parseInt(req.getParameter("id"));
         System.out.println(id);
         if(req.getParameter("action").equals("edit")){
             User user = userServiceImpl.getUserById(id);
