+++ src/com/bsu/by/komodey/controllers/ModifyUserController.java
@@ -9,13 +9,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
 @Controller
 public class ModifyUserController {
     private UserService userServiceImpl =
@@ -42,7 +35,7 @@ public class ModifyUserController {
                         String phone,
                         String check,
                         String id){
-        long idL = Long.parseLong(id);
+        int idL = Integer.parseInt(id);
         String role;
         if(check==null){
             role = "user";
